page/requirements.txt
Flask==1.1.2
requests==2.25.1
algoliasearch==2.4.0
beautifulsoup4==4.9.3
firebase_admin==4.5.1
flask_login==0.5.0
google_api_python_client==1.12.8
google_auth_oauthlib==0.4.2
httplib2==0.19.0
line_profiler==3.1.0
livereload==2.6.3
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.14.0
python-dotenv==0.15.0
selenium==3.141.0
webdriver_manager==3.3.0
